Former-commit-id: 4d502d15a9c450820869ece5b64cc14457d83370

+++ Tools/ResourceEditor/Classes/SceneEditor/HeightmapNode.h
@@ -47,7 +47,7 @@ protected:
     
     RenderDataObject *renderData;
 
-    Vector<uint8> hmap;
+    Vector<uint16> hmap;
     
     Vector<float32> verts;
     Vector<float32> colors;
